refactor(web): shrink CSV artifact preview after ponytail review
Fold the panel hook back, drop pass-through wrappers, and keep the
CSV helpers private with structuredClone for row copies.
